+++ tests/doc_test/doc_global_options/conf.py
@@ -48,7 +48,7 @@ needs_global_options = {
     "global_5": ("STATUS_CLOSED", 'status == "closed"', "STATUS_UNKNOWN"),
 }
 
-plantuml = "java -jar %s" % os.path.join(os.path.dirname(__file__), "..", "..", "..", "docs", "utils", "plantuml.jar")
+plantuml = "java -jar %s" % os.path.join(os.path.dirname(__file__), "..", "utils", "plantuml.jar")
 plantuml_output_format = "svg"
 
 # Add any paths that contain templates here, relative to this directory.