+++ tests/test_api_usage_in_extension.py
@@ -2,7 +2,7 @@ import imp
 import sys
 from pathlib import Path
 
-from sphinx_testing import with_app
+import pytest
 
 dummy_code = """
 def setup(app):
@@ -22,8 +22,10 @@ exec(dummy_code, dummy_extension.__dict__)
 sys.modules["dummy_extension.dummy"] = dummy_extension
 
 
-@with_app(buildername="html", srcdir="doc_test/api_doc")
-def test_api_configuration(app, status, warning):
+@pytest.mark.parametrize("test_app", [{"buildername": "html", "srcdir": "doc_test/api_doc"}], indirect=True)
+def test_api_configuration(test_app):
+    app = test_app
+
     app.builder.build_all()
     html = Path(app.outdir, "index.html").read_text()
     assert html is not None