+++ tests/test_broken_statuses.py
@@ -1,13 +1,13 @@
-from nose.tools import raises
-from sphinx_testing import with_app
+import pytest
 
 from sphinxcontrib.needs.api.need import NeedsStatusNotAllowed
 
 
-@raises(NeedsStatusNotAllowed)
-@with_app(buildername="html", srcdir="doc_test/broken_statuses")
-def test_doc_build_html(app, status, warning):
-    app.build()
-    html = (app.outdir / "index.html").read_text()
-    assert "<h1>BROKEN DOCUMENT" in html
-    assert "SP_TOO_002" in html
+@pytest.mark.parametrize("test_app", [{"buildername": "html", "srcdir": "doc_test/broken_statuses"}], indirect=True)
+def test_doc_build_html(test_app):
+    with pytest.raises(NeedsStatusNotAllowed):
+        app = test_app
+        app.build()
+        html = (app.outdir / "index.html").read_text()
+        assert "<h1>BROKEN DOCUMENT" in html
+        assert "SP_TOO_002" in html