+++ tests/test_external.py
@@ -1,11 +1,12 @@
 import json
 from pathlib import Path
 
-from sphinx_testing import with_app
+import pytest
 
 
-@with_app(buildername="html", srcdir="doc_test/external_doc")  # , warningiserror=True)
-def test_external_html(app, status, warning):
+@pytest.mark.parametrize("test_app", [{"buildername": "html", "srcdir": "doc_test/external_doc"}], indirect=True)
+def test_external_html(test_app):
+    app = test_app
     app.build()
     html = Path(app.outdir, "index.html").read_text()
     assert (
@@ -14,8 +15,9 @@ def test_external_html(app, status, warning):
     )
 
 
-@with_app(buildername="needs", srcdir="doc_test/external_doc")  # , warningiserror=True)
-def test_external_json(app, status, warning):
+@pytest.mark.parametrize("test_app", [{"buildername": "needs", "srcdir": "doc_test/external_doc"}], indirect=True)
+def test_external_json(test_app):
+    app = test_app
     app.build()
     json_data = Path(app.outdir, "needs.json").read_text()
     needs = json.loads(json_data)