+++ tests/test_needs_builder.py
@@ -1,11 +1,12 @@
 import json
 from pathlib import Path
 
-from sphinx_testing import with_app
+import pytest
 
 
-@with_app(buildername="needs", srcdir="doc_test/doc_needs_builder")
-def test_doc_needs_builder(app, status, warning):
+@pytest.mark.parametrize("test_app", [{"buildername": "needs", "srcdir": "doc_test/doc_needs_builder"}], indirect=True)
+def test_doc_needs_builder(test_app):
+    app = test_app
     app.build()
 
     needs_json = Path(app.outdir, "needs.json")
@@ -21,12 +22,16 @@ def test_doc_needs_builder(app, status, warning):
     assert needs_list["versions"]["2.0"]["needs"]["TEST_01"]
 
 
-@with_app(buildername="needs", srcdir="doc_test/doc_needs_builder_negative_tests")
-def test_doc_needs_build_without_needs_file(app, status, warning):
+@pytest.mark.parametrize(
+    "test_app", [{"buildername": "needs", "srcdir": "doc_test/doc_needs_builder_negative_tests"}], indirect=True
+)
+def test_doc_needs_build_without_needs_file(test_app):
     import os
     import subprocess
 
-    srcdir = "doc_test/doc_needs_builder_negative_tests"
+    app = test_app
+
+    srcdir = Path(app.srcdir)
     out_dir = os.path.join(srcdir, "_build")
 
     out = subprocess.run(["sphinx-build", "-b", "needs", srcdir, out_dir], capture_output=True)
@@ -34,13 +39,16 @@ def test_doc_needs_build_without_needs_file(app, status, warning):
     assert "needs.json found, but will not be used because needs_file not configured." in out.stdout.decode("utf-8")
 
 
-@with_app(buildername="needs", srcdir="../docs")
-def test_needs_official_doc(app, status, warning):
+@pytest.mark.parametrize("test_app", [{"buildername": "needs", "srcdir": "../docs"}], indirect=True)
+def test_needs_official_doc(test_app):
+    app = test_app
     app.build()
 
 
-@with_app(buildername="html", srcdir="doc_test/doc_needs_builder_parallel")
-def test_needs_html_and_json(app, status, warning):
+@pytest.mark.parametrize(
+    "test_app", [{"buildername": "html", "srcdir": "doc_test/doc_needs_builder_parallel"}], indirect=True
+)
+def test_needs_html_and_json(test_app):
     """
     Build html output and needs.json in one sphinx-build
     """
@@ -48,12 +56,13 @@ def test_needs_html_and_json(app, status, warning):
     import os
     import subprocess
 
+    app = test_app
     app.build()
 
     needs_json_path = os.path.join(app.outdir, "needs.json")
     assert os.path.exists(needs_json_path)
 
-    srcdir = "doc_test/doc_needs_builder_parallel"
+    srcdir = app.srcdir
     build_dir = os.path.join(app.outdir, "../needs")
     subprocess.run(["sphinx-build", "-b", "needs", srcdir, build_dir], capture_output=True)
     needs_json_path_2 = os.path.join(build_dir, "needs.json")