+++ tests/test_needs_filter_data.py
@@ -1,10 +1,13 @@
 from pathlib import Path
 
-from sphinx_testing import with_app
+import pytest
 
 
-@with_app(buildername="html", srcdir="doc_test/doc_needs_filter_data")
-def test_doc_needs_filter_data_html(app, status, warning):
+@pytest.mark.parametrize(
+    "test_app", [{"buildername": "html", "srcdir": "doc_test/doc_needs_filter_data"}], indirect=True
+)
+def test_doc_needs_filter_data_html(test_app):
+    app = test_app
     app.build()
     index_html = Path(app.outdir, "index.html").read_text()
 
@@ -45,6 +48,7 @@ def test_doc_needs_filter_data_html(app, status, warning):
     )
 
     # check needs_warnings works
+    warning = app._warning
     warnings = warning.getvalue()
 
     # check warnings contents
@@ -53,8 +57,11 @@ def test_doc_needs_filter_data_html(app, status, warning):
     assert "used filter: variant != current_variant" in warnings
 
 
-@with_app(buildername="html", srcdir="doc_test/doc_needs_filter_data")
-def test_doc_needs_filter_code(app, status, warning):
+@pytest.mark.parametrize(
+    "test_app", [{"buildername": "html", "srcdir": "doc_test/doc_needs_filter_data"}], indirect=True
+)
+def test_doc_needs_filter_code(test_app):
+    app = test_app
     app.build()
     code_html = Path(app.outdir, "filter_code.html").read_text()