+++ tests/test_needs_warning.py
@@ -1,11 +1,15 @@
-from sphinx_testing import with_app
+from pathlib import Path
+
+import pytest
 
 
-@with_app(buildername="html", srcdir="doc_test/doc_needs_warnings")  # , warningiserror=True)
-def test_needs_warnings(app, status, warning):
+@pytest.mark.parametrize("test_app", [{"buildername": "html", "srcdir": "doc_test/doc_needs_warnings"}], indirect=True)
+def test_needs_warnings(test_app):
+    app = test_app
     app.build()
 
     # stdout warnings
+    warning = app._warning
     warnings = warning.getvalue()
 
     # check multiple warning registration
@@ -31,13 +35,16 @@ def test_needs_warnings(app, status, warning):
     assert "EXT_TEST_01" not in warnings
 
 
-@with_app(buildername="html", srcdir="doc_test/doc_needs_warnings_return_status_code")
-def test_needs_warnings_return_status_code(app, status, warning):
-    import os
+@pytest.mark.parametrize(
+    "test_app", [{"buildername": "html", "srcdir": "doc_test/doc_needs_warnings_return_status_code"}], indirect=True
+)
+def test_needs_warnings_return_status_code(test_app):
     import subprocess
 
-    srcdir = "doc_test/doc_needs_warnings_return_status_code"
-    out_dir = os.path.join(srcdir, "_build")
+    app = test_app
+
+    srcdir = Path(app.srcdir)
+    out_dir = srcdir / "_build"
 
     # Check return code when "-W --keep-going" not used
     out_normal = subprocess.run(["sphinx-build", "-M", "html", srcdir, out_dir], capture_output=True)