+++ tests/test_role_need_max_title_length.py
@@ -1,13 +1,19 @@
 import os
 from pathlib import Path
 
-from sphinx_testing import with_app
+import pytest
 
 
-@with_app(buildername="html", srcdir="doc_test/doc_role_need_max_title_length_unlimited")
-def test_max_title_length_unlimited(app, status, warning):
+@pytest.mark.parametrize(
+    "test_app",
+    [{"buildername": "html", "srcdir": "doc_test/doc_role_need_max_title_length_unlimited"}],
+    indirect=True,
+)
+def test_max_title_length_unlimited(test_app):
 
     os.environ["MAX_TITLE_LENGTH"] = "-1"
+
+    app = test_app
     app.build()
     html = Path(app.outdir, "index.html").read_text()
     assert "ROLE NEED TEMPLATE" in html
@@ -17,10 +23,14 @@ def test_max_title_length_unlimited(app, status, warning):
     )
 
 
-@with_app(buildername="html", srcdir="doc_test/doc_role_need_max_title_length")
-def test_max_title_length_10(app, status, warning):
+@pytest.mark.parametrize(
+    "test_app", [{"buildername": "html", "srcdir": "doc_test/doc_role_need_max_title_length"}], indirect=True
+)
+def test_max_title_length_10(test_app):
 
     os.environ["MAX_TITLE_LENGTH"] = "10"
+
+    app = test_app
     app.build()
     html = Path(app.outdir, "index.html").read_text()
     assert "ROLE NEED TEMPLATE" in html