+++ tests/test_role_need_template.py
@@ -1,11 +1,13 @@
 from pathlib import Path
 
-from sphinx_testing import with_app
+import pytest
 
 
-@with_app(buildername="html", srcdir="doc_test/doc_role_need_template")
-def test_doc_build_html(app, status, warning):
-
+@pytest.mark.parametrize(
+    "test_app", [{"buildername": "html", "srcdir": "doc_test/doc_role_need_template"}], indirect=True
+)
+def test_doc_build_html(test_app):
+    app = test_app
     app.build()
     html = Path(app.outdir, "index.html").read_text()
     assert "ROLE NEED TEMPLATE" in html