+++ tests/test_title_optional.py
@@ -1,8 +1,8 @@
 import sys
 from pathlib import Path
 
+import pytest
 import sphinx
-from sphinx_testing import with_app
 
 from tests.util import extract_needs_from_html
 
@@ -26,8 +26,8 @@ class HtmlNeed:
         return title[0].text if title is not None else None  # title[0] aims to the span_data element
 
 
-@with_app(buildername="html", srcdir="doc_test/title_optional")
-def test_title_optional_scenarios(app, status, warning):
+@pytest.mark.parametrize("test_app", [{"buildername": "html", "srcdir": "doc_test/title_optional"}], indirect=True)
+def test_title_optional_scenarios(test_app):
 
     # Somehow the xml-tree in extract_needs_from_html() works not correctly with py37 and specific
     # extracts, which are needed for sphinx >3.0 only.
@@ -35,6 +35,7 @@ def test_title_optional_scenarios(app, status, warning):
     if sys.version_info[0] == 3 and sys.version_info[1] == 7 and sphinx.version_info[0] >= 3:
         return True
 
+    app = test_app
     app.build()
 
     html = Path(app.outdir, "index.html").read_text()